Fix reading as incorrect float instead of short

+++ CheesewheelCollab/Assets/Source/Networking/NetworkGameManager.cs
@@ -211,7 +211,7 @@ namespace Source.Networking
             Time = reader.GetFloat();
             for (var i = 0; i < Samples.Length; i++)
             {
-                Samples[i] = (short)MathUtility.Remap(reader.GetFloat(), short.MinValue, short.MaxValue, -1, 1);
+                Samples[i] = (short)MathUtility.Remap(reader.GetShort(), short.MinValue, short.MaxValue, -1, 1);
             }
         }
     }
